[api] Refactoring FeatureHelper
- Moving API specific code from app/helpers/feature_helper.rb
to the API Controller's features.rb file.
